Fix playground when running in minikube (#279)
The playground UI would ask the playground service to fetch sources from `browser_address/sources/<function-name>` - e.g. `192.168.64.4:32050/sources/echo.go`. The problem is that minikube (and probably Virtualbox) do not have the routes in place to allow a container to access that URL. 

This works fine in Kubernetes on AWS given that `browser_address` is the elastic IP and the container can reach this just fine.