Consolidate processor.yaml / build.yaml into function.yaml (#290)
This PR re-arranges how configuration is both declared and injected into a processor. An important terminology change is also introduced - "event source" is now "trigger" both in the code and this description. 

Prior to this PR, the user had two ways of defining the configuration of a processor (where configuration contains things like runtime, handler, triggers, data bindings, and many others): 
* Through defining a processor.yaml - the raw configuration file used by the processor
* By populating a function custom resource in kubernetes would create a processor.yaml of sorts (not all fields existed in the function CR).

The user could also specify build-time parameters through build.yaml. 

After this PR the user inputs a single configuration. This single configuration can be in the form of: 
* A function.yaml file 
* A kubernetes function CR 
* A kubernetes function spec file (will be introduced in a different PR - it's just a matter of parsing + lots of tests)
 
The thing is the 3 above *share the same structure* and contain everything about the function. An example configuration can be: 

```
runtime: python
handler: some_module:handler

triggers:
  rmq:
    kind: rabbit-mq
    url: amqp://guest:guest@172.17.0.1:5672
    attributes:
      exchangeName: nuclio.rabbitmq_trigger_test
      queueName: test_queue

dataBindings:
  db0:
    class: v3io
    url: http://foo:1000/1024
    secret: whatever
  db1:
    class: v3io
    url: http://foo:1000/1025
    secret: whatever

build:
  baseImageName: ubuntu:16.04
  commands:
    - pip install simplejson
```

The processor still expects a processor.yaml to exist and since runtime configuration is a requirement, the processor.yaml is generated on deploy and injected to the pod as a configmap (via a volume in local platform).

Besides this change - lots of consolidation around the platform was made.