Delete existing functions by name in local platform (#654)
Following a recent PR that set the container name to `namespace-name` in the local platform, if a function failed to deploy subsequent deploys would fail. This is because the local platform looked for existing functions via running containers with given labels. 

This PR changes the search for existing functions to be based on the container name.