User can control the default ingress rule (#700)
This PR introduces the concept of `defaultIngressPattern`, configurable as part of the HTTP trigger. Prior to this change, a `/function-name/function-version` ingress rule would always be created. In some cases, users do not want to expose functions via ingresses or perhaps control this pattern. This is possible with this PR. The user can:
1. Do nothing, and behavior will be backwards compatible
2. Specify an empty string for `defaultIngressPattern` - in which case no ingress rule will be created
3. Specify a pattern for `defaultIngressPattern` to control the default ingress rule (more on this in the changed docs)

In addition, a few bugs were fixed in this area:
1. Ingress rules were incorrectly created since 0.2.8. They would be in the form of `/function-name/` rather than `/function-name/function-version`
2. Ingresses can now be updated

Fixes #697.