Allow triggers to wait for an available worker rather than return error immediately (#910)
Before this PR, if an event arrived at a trigger and no worker was available to handle it - the trigger would receive and return an error (e.g. 503 in HTTP). This PR introduces a new trigger configuration with a snappy name: workerAvailabilityTimeoutMilliseconds that allows specifying a timeout.

If this value is left unset or set to 0 - the behavior of the trigger is unchanged (including no performance hit). If this value is set, the trigger will wait for a worker to become available for this duration of time (with a certain amount of performance hit - to be determined by benchmarking).