Update NATS trigger to use queue subscriptions (#944)
The NATS trigger currently creates a new standard subscription for each processor instance. As a result, every replica of a deployed function receives a notification for every message published to the configured topic (i.e. every message is processed 'n' times). This essentially eliminates the benefits of distributed event handling (with the exception of some level of increased certainty that an event will be handled at least once).

NATS supports an additional round-robin messaging type that allows subscribers to specify a (worker group) queue name along with a topic pattern when subscribing. The queue acts as a single subscriber, and messages are load-balanced across members of the worker group. See [NATS Docs](https://nats.io/documentation/writing_applications/subscribing/)

This PR updates the NATS trigger to enable processor instances associated with the same trigger definition to join the same queue. A queue name may be supplied or generated automatically.