Fix build from URL (#1291)
Providing a URL to a non-Go function (e.g. http://www.foo.com/my-function.py) failed. This is because a recent change was made to rename the function after download. In addition, the integration test to verify downloading files from URLs was not proper.

This PR fixes downloading files from remote URLs (doesn't rename them, yet maintains the behavior of renaming archivees) and fixes the integration test. It also refactors a long function (`resolveFunctionPath`) to be easier to read.