Support for Kafka partition allocators (#1477)
A few commits before this change, a single Kafka partition was handled in a single worker. This meant users needed to specify maxWorkers = max number of partitions a replica can handle. If for example you had a topic with 64 partitions, you needed to specify maxWorkers = 64 to allow a case where a single replica handles all the partitions (or you would fail deploy due to worker allocation failure). This is problematic because if you use scaling or more than one replica, you will have lots of wasted workers (e.g. for 64 partitions and 8 replicas w/maxWorker = 64, you have a total of 512 workers across all replicas). For runtimes like Python, this consumes quite a bit of memory.

After that, a small change was made so that when a messages arrives at a partition it allocates a worker from the pool, handles and releases the worker. This means that users could provide any number of workers (e.g. much smaller than topic partitions) and the messages could be handled in the workers according to the worker availability. However, this means that there is no more guarantee that a single partition will be handled by the same worker across the lifetime of the replica. For stateless functions that's fine, but stateful functions may suffer.

This change introduces a new mode: users can specify any number of workers, and these are statically divided among the partitions. For example, if you have 6 partitions and 2 workers - the first 3 partitions will ALWAYS be handled by the first worker and the last 3 partitions will ALWAYS be handled by the last 3 partitions. This means that throughput may suffer if the partitions are not equally loaded, but is a good balance between requiring 1:1 partition:worker (the first solution) and a worker pool (the second solution).

It is possible to configure the kafka trigger worker allocation mode to choose between pool (the second solution) and static (the new solution). The default is the pooled mode.

In addition to the above, configuration for kafka can be controlled through annotations to allow users to specify this before it's available in the UI.