UI: Omit project name from error message on single project action
When multiple projects are selected and a single action is taken upon
all of them it makes sense to have each project's name before its
corresponding error message so the user knows which error message is
relevant to which project.
When an action is taken on a single project the project name is
superfluous.