[Golang] Fix go runtime on low-latency setup (#3510)
There was an issue with the low-latency test in the Go runtime.

The problem was that we sometimes attempted to write to the channel when
the reader wasn’t ready yet. In this case, a blocking write is
acceptable because if the runtime is terminated, the channel will be
closed, causing the writing goroutine to panic and handle the panic
accordingly (which is covered in the Go tests).

Additionally, I added a test that reproduces the issue. It only occurred
after 10,000 iterations, so I set the test to run for 10,000 iterations.
While that may seem like a lot, it doesn’t take much time and serves as
a reliable test to ensure the stability of the Go runtime.

UPD: test fails because of race condition in metrics, working on it