[Processor] Remove spammy logs from async wrapper (#3554)
Too many logs from too many async tasks at the same time may lead to the
problem where socket buffer is full. This problem should be handled from
logger side (nuclio-sdk-py) as well. In previous implementation we never
saw this problem because the general load on logging in sockets was
incomparable with the current one.