[Processor] New restart handling logic (#3553)
The idea is to move the logic that determines whether a restart is
needed to the worker.
The timeout mechanism simply receives a boolean response and restarts
the worker/processor if necessary.

Async restart flow:

1. If event processing timeout-ed, connection is marked as
`restartRequired`
2. when releasing (btw, releasing now happens in a separate goroutine
and doesn't affect execution time anymore), check connection status. If
restart required, create new connection and replace the old one with it.
If creation fails after 3 retries, mark whole connection allocator as
`restartRequired`.
3. when connection allocator is in `restartRequired` state, ready
connections are still available for allocation.
4. meanwhile, watcher() process is checking if runtime requires restart,
runtime checks if connection manager requires restart and if it does,
runtime is restarted.

Other changes:

Introduced connection allocator release logic and improved error
handling (including restart support).
Added support for socket-level restarts on timeout. Now, for all RPC
runtimes, the event processor can detect timeouts directly, allowing it
to respond to the user immediately. This eliminates the need to wait for
a background task to trigger the restart process. From the restart
perspective, the behavior remains the same—the wrapper is still
restarted by the timeout watcher.

https://iguazio.atlassian.net/browse/NUC-385
https://iguazio.atlassian.net/browse/NUC-400