[Dependencies] Change go version to 1.23.8 (#3588)
Change go version to 1.23.8, also setting images to have a patch
version.

During security check we noticed misalignment that was caused by 1.23.0
in go.mod and 1.23.7 version of the go image itself. In order to have
clarity of which version we use, it was decided to always have a patch
version specified.