[Processor] Implement non-blocking pool allocator (#3608)
Jira - https://iguazio.atlassian.net/browse/NUC-410

The base branch is
[feature/nuc-201-phase2](https://github.com/nuclio/nuclio/tree/feature/nuc-201-phase2),
so we will merge to develop only when everything is ready.

Implements a new non-blocking pool allocator that allocates objects
using a round-robin strategy. The round-robin mechanism is driven by an
atomic index, which is incremented with each allocation. Additionally,
the shared pool logic has been refactored into an abstract base class,
`abstractPoolAllocator`.

The `BenchmarkParallelAllocation` test has also been made generic for
different allocator implementations. This allows us to pass in a
constructor and benchmark any allocator, which is valuable since the
allocator component is a critical part of overall performance.