[Processor] Check status when allocating from non-blocking pool (#3621)
Improves the allocation method in the non-blocking pool so that it
checks the state of the allocated object. If the object is not ready, it
will retry N times, where N is the number of objects in the allocator.

Also, replaced `Equal` with `Same` in allocator test. 
When mocks like MockEventProcessor are compared, identical internal
state can cause them to be seen as equal even if they are distinct
instances.
What needs to be verified in the tests is whether two different
instances are returned, not just objects with matching fields.
With Same, a comparison based on memory address is performed, ensuring
that true object identity is validated.
For this reason, Same (or NotSame) is preferred over Equal (or NotEqual)
when verifying allocation behavior with mocks.

Jira - https://iguazio.atlassian.net/browse/NUC-413