[Trigger] Add NATS Jetstream Trigger (#3649)
This feature is about extending the current NATS trigger capabilities by
adding a Jetstream option. The main advantages of using Jetstream over
NATS pub/sub is the ability to store and replay messages. If for some
reason there are no available function replica, messages sent to NATS
are not lost and will be forwarded to a function replica as soon as one
is available.

Currently this trigger is not reflected on the dashboard. Please let me
know if I need to create a separate MR on the dashboard-controls
repository.