[OPA] Use opa-client package (#3692)
Use the new https://github.com/nuclio/opa-client package as the HTTP OPA
client.
This package was extracted from the `nuclio/nuclio` repo so it can be
used by other projects as well, so it basically has the same
implementation.

The only difference is that now both Query functions (Filter single &
multiple resources) get a `ctx` argument.

The internal `opa` package is left to handle constants and helper
functions that Nuclio uses, but are not specific to the client's
implementation.

A new `Mode` was added to the OPA section in the platform config, as a
prep for working with Iguazio V4.
In V4, the resources are namespaced under either `/resources` or
`/mgmt`, so this prefix is added according to this mode.
This is backwards compatible, as if no mode is provided - no prefix is
added.

https://iguazio.atlassian.net/browse/NUC-507