[Controller] Enrich and Validate service account (#3711)
Since service account enrichment and validation is needed in both
controller and dashboard, this PR follows the logic of `NodeSelectors`
implementation which we enrich on controller level only and write
enriched value into `function.Status`

The logic of enrichment and validation of the function's SA (excluding
Kaniko for now - will be done in a separate PR):

1. On function creation, dashboard validates that if function.SA is set,
it is allowed. If not, fails fast
2. If 1 step was successful, then dashboard creates CRD successfully,
and controller steps in
3. Controller on function create/update enriches the value as we usually
do (function -> project-> platform) and then validates it again. If SA
is not allowed, fails.

All the SA-related code was moved to `common/k8s.go` to be able to share
it between controller and dashboard, all functions are covered in the
tests.

Jira - https://iguazio.atlassian.net/browse/NUC-519