[Controller] Check for scaling happened within last scalingTimeout even if CurrentReplicas is equal to DesiredReplicas (#3722)
### Description  
This PR fixes an issue where a function was incorrectly marked as
unhealthy even though it was in the process of scaling.

The `FunctionMonitor` determines readiness by first checking
`isAvailable()`. If the function is not yet available, it checks whether
it is currently scaling using the isScaling() method, which inspects the
HPA status.

However, the previous logic missed one scenario: when pods had already
been created but were not yet ready. In such cases, the condition
`hpa.Status.CurrentReplicas != hpa.Status.DesiredReplicas `was bypassed,
since `CurrentReplicas ` includes all pods (not just the ready ones). As
a result, the function was considered not scaling, which led the monitor
to mark it as unhealthy.

This PR adjusts the logic to properly account for this intermediate
state, ensuring that the function is recognized as scaling when
appropriate and not prematurely marked as unhealthy.

### Testing  
It is quite complicated to test this change, as bug isn't always
reproduces, so it will be tested by QA.

### References
- Ticket link: https://iguazio.atlassian.net/browse/NUC-288