[Builder] Take only image name when resolving base image with custom registry (#3724)
### Description  
PR #3710 changed the base images location, so we take all the images
from gcr instead of docker hub. However, it turned out that this breaks
existing iguazio system because when resolving the base image on the
system with custom registry, Nuclio constructs the final base image by
prepending the custom registry and stripping the original registry from
the hardcoded image to extract just the image name. In order to keep BC
it was decided that from now on we will get only the name (not a full
path in registry) and construct the image path.

Also, aligned java image usage with existing image on iguazio.

### Testing  
- [x] I have tested the changes in this PR

<img width="2475" height="638" alt="image"
src="https://github.com/user-attachments/assets/6a07cfbf-a48e-433e-977d-3deffb3b74fd"
/>

dotnet has issue only because of old onbuild on 7.0 version, but image
was built correctly anyway.