[Dashboard] Do not concatenate run registry if image already has registry (#3745)
### Description  
Refactored logic for enriching the image with the run registry.
Previously, the code checked whether the image was already prefixed with
the run registry using a string prefix check. This has been replaced
with a more robust check using the shared ImageHasRegistry method, which
determines whether the image includes a registry component at all.


### Testing  
- [x] I have tested the changes in this PR
Unit tests

### References
- Ticket link: jira https://iguazio.atlassian.net/browse/NUC-552