[Dashboard] Add connectionAvailabilityTimeout to async config (#3766)
### 📝 Description
Connection allocation timeout used to be set to 0 by default which
caused the allocation error when all the existing connections are in
use. In order to fix that and align the approach with worker allocation,
a new parameter was added to async config -
`connectionAvailabilityTimeout`. By default it's `10s` - the same as
worker allocation timeout. Worker allocation timeout isn't reused here
to make it more clear what each value means.

---

### 🛠️ Changes Made
Value is enriched on the dashboard level.

---

### ✅ Checklist
- [x] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
unit test + tested on vmdev

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/NUC-560
---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No