[Tests] Improve kafka trigger tests (#3821)
### 📝 Description
Adds a test with configurable number of draining cycles which are
created by deploying new functions on every cycle. This test ensures
that all the messages were processed and committed.

---

### 🛠️ Changes Made
* add a test for combination of different streaming features
* fixed bug https://iguazio.atlassian.net/browse/NUC-609. We used to
terminate functions in a local platform using `docker rm -f` which is
basically a `SIGKILL` which didn't allow for graceful termination. Now
when deleting function, nuclio does `docker stop -t 30` which gives 30s
to function to terminate (the same time as kubernetes has). And then we
do `rm -f` just to make sure that function is removed no matter what.

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/NUC-608,
https://iguazio.atlassian.net/browse/NUC-605,
https://iguazio.atlassian.net/browse/NUC-609
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---------

Co-authored-by: TomerShor <90552140+TomerShor@users.noreply.github.com>