[Processor] Return a copy of statistics objects during metrics gathering (#3907)
### 📝 Description
Follows the same pattern as with allocation statistics, where with
statistics itself we use safe atomic, but when returning stats to
gatherers, we copy an object to an unsafe version (with normal uints,
not with atomics) and return it. This way, we still are on the safe side
and thread-safe during statistics calculation, but thread-unsafe (which
is ok) during gathering, because metrics gathering is synchronous


---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
tested on vmdev

---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/NUC-658
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [ ] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<!-- Anything else reviewers should know (follow-up tasks, known issues,
affected areas etc.). -->
<!-- ### 📸 Screenshots / Logs -->