[Dashboard] Allow nuclio logs header in invocation endpoint (#3924)
### 📝 Description
The reason why it didn't work from UI was the fact the UI uses
dashboard's invocation endpoint, which filtered out all Nuclio custom
headers. This PR adds logs header to list of allowed headers and aligned
the allowed headers list with the one we have in `ExposedHeaders`, so
that if we need to modify it, we can modify it in a single place.

---

### ✅ Checklist
- [ ] I updated the documentation (if applicable)
- [x] I have tested the changes in this PR

---

### 🧪 Testing
Tested on vmdev, see the attached screenshor
---

### 🔗 References
- Ticket link: https://iguazio.atlassian.net/browse/NUC-670
- Design docs links:
- External links:

---

### 🚨 Breaking Changes?

- [ ] Yes (explain below)
- [x] No

<!-- If yes, describe what needs to be changed downstream: -->

---

### 🔍️ Additional Notes
<img width="833" height="1046" alt="image"
src="https://github.com/user-attachments/assets/88bab99a-1068-4402-8e59-fde8e058922b"
/>

Co-authored-by: TomerShor <90552140+TomerShor@users.noreply.github.com>